Support for Maven Archives.
A Maven ARchive (.mar file) is just like a Web ARchive (.war file),
except that the Maven jar dependencies in WEB-INF/lib are replaced
with entries in WEB-INF/lib/maven-dependencies.json. This drastically
reduces the size of the archive. At deploy time, this CloudFoundry build
pack can be used to reconstitute the original archive.